funding-bot/requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
hyperliquid-python-sdk>=0.1.0
eth-account>=0.10.0
